+++ sldl_compiler/__init__.py
@@ -1,3 +1,3 @@
-"""SLDL v1.0.1 Python compiler."""
+"""SLDL v1.0.2 Python compiler."""
 
-__version__ = "1.0.1"
+__version__ = "1.0.2"